Remove separate helpers module; add to requirements; minor cleanup
