+++ requirements.txt
@@ -1,2 +1,3 @@
 napalm_base>=0.23.2
 netmiko>=1.4.0
+netaddr
